# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: microHelium API
  description: >
    HTTP API for microHelium, a Laravel-based ICPC-style contest judge.
    Covers every route currently registered in routes/api.php. This spec is
    hand-maintained against the Api\* controllers under
    app/Http/Controllers/Api -- tests/Feature/Api/OpenApiSpecTest.php checks
    every registered API route has a matching path documented here, so a new
    endpoint without a spec entry fails the suite.
  version: "1.0.0"
servers:
  - url: /api
security:
  - sanctum: []
...
paths:
...
  /clarifications:
    get:
      tags: [Clarifications]
      summary: List clarifications
      description: >
        Teams only see their own clarifications plus anything with status
        broadcast_site or broadcast_all; admins/judges see everything.
        Despite the name, broadcast_site is currently NOT filtered to the
        answering judge's site -- it is visible contest-wide, same as
        broadcast_all (see app/Http/Controllers/Api/ClarificationController.php
        ::index(), tracked as a known inaccuracy vs. the status name).
      parameters:
        - $ref: "#/components/parameters/ContestIdQuery"
      responses:
        "200":
          description: Paginated list of clarifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedClarifications"
    post:
      tags: [Clarifications]
      summary: Ask a clarification question
      description: Fails with 422 if the contest is not currently running.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contest_id, question]
              properties:
                contest_id: { type: integer }
                problem_id: { type: integer, nullable: true }
                question: { type: string, maxLength: 2000 }
      responses:
        "201":
          description: Clarification submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clarification"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /clarifications/{clarification}/answer:
    parameters:
      - $ref: "#/components/parameters/ClarificationId"
    put:
      tags: [Clarifications]
      summary: Answer a clarification
      description: >
        Intended for judges/admins, but ClarificationController::answer()
        currently enforces no such check beyond auth:sanctum -- any
        authenticated user, including a team, can answer any clarification.
        Tracked as a known authorization gap, not a documentation error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answer]
              properties:
                answer: { type: string, maxLength: 2000 }
                broadcast:
                  type: string
                  enum: [none, site, all]
                  nullable: true
      responses:
        "200":
          description: Answered clarification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clarification"
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      description: "Laravel Sanctum personal-access token, sent as `Authorization: Bearer <token>`."
    webcastCredential:
      type: http
      scheme: bearer
      description: >
        Issue #44 webcast credential secret, sent as `Authorization: Bearer
        <secret>`. Verified against webcast_credentials.token_hash by
        App\Http\Middleware\AuthenticateWebcastCredential -- entirely
        separate from Sanctum and from any `user_type`.
  parameters:
...
    ContestIdQuery:
      name: contest_id
      in: query
      required: false
      schema: { type: integer }
...
    ClarificationId:
      name: clarification
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
  schemas:
...
    Clarification:
      type: object
      properties:
        id: { type: integer }
        contest_id: { type: integer }
        site_id: { type: integer }
        user_id: { type: integer }
        problem_id: { type: integer, nullable: true }
        clarification_number: { type: integer }
        question: { type: string }
        answer: { type: string, nullable: true }
        status:
          type: string
          enum: [pending, answered, broadcast_site, broadcast_all, deleted]
        contest_time: { type: integer }
        answered_time: { type: integer, nullable: true }
        judge_id: { type: integer, nullable: true }
...
    PaginatedClarifications:
      allOf:
        - $ref: "#/components/schemas/PaginationMeta"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Clarification" }
    PaginationMeta:
      type: object
      properties:
        current_page: { type: integer }
        last_page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }